+++ packages/gatsby/src/query/query-compiler.js
@@ -430,7 +430,8 @@ const processDefinitions = ({
 const determineUsedFragmentsForDefinition = (
   definition,
   definitionsByName,
-  fragmentsUsedByFragment
+  fragmentsUsedByFragment,
+  visitedFragments = new Set()
 ) => {
   const { def, name, isFragment, filePath } = definition
   const cachedUsedFragments = fragmentsUsedByFragment.get(name)
@@ -444,6 +445,10 @@ const determineUsedFragmentsForDefinition = (
         const name = node.name.value
         const fragmentDefinition = definitionsByName.get(name)
         if (fragmentDefinition) {
+          if (visitedFragments.has(name)) {
+            return
+          }
+          visitedFragments.add(name)
           usedFragments.add(name)
           const {
             usedFragments: usedFragmentsForFragment,
@@ -451,7 +456,8 @@ const determineUsedFragmentsForDefinition = (
           } = determineUsedFragmentsForDefinition(
             fragmentDefinition,
             definitionsByName,
-            fragmentsUsedByFragment
+            fragmentsUsedByFragment,
+            visitedFragments
           )
           usedFragmentsForFragment.forEach(fragmentName =>
             usedFragments.add(fragmentName)
@@ -494,10 +500,13 @@ const addExtraFields = (document, schema) => {
         // Entering a field of the current selection-set:
         //   mark which fields already exist in this selection set to avoid duplicates
         const context = contextStack[contextStack.length - 1]
-        if (node.name.value === `__typename`) {
+        if (
+          node.name.value === `__typename` ||
+          node?.alias?.value === `__typename`
+        ) {
           context.hasTypename = true
         }
-        if (node.name.value === `id`) {
+        if (node.name.value === `id` || node?.alias?.value === `id`) {
           context.hasId = true
         }
       },